requirements.txt
pyserial == 3.5
ascon == 0.0.9